testing/requirements.txt
pytest
pandas
selenium
webdriver-manager
python-dotenv